Refactor work order management features
- Updated import paths for message translations and components to improve modularity.
- Consolidated work order details into a nested MessagesGroup for better organization.
- Removed deprecated components and views related to work orders to streamline the codebase.
- Enhanced the WorkOrdersView to reflect the new structure and maintain functionality.
- Refactored types for improved clarity and consistency in work order data handling.
