+++ modules/projects/settings/views/main-view/tab-views/work-orders/actions/dialogs/ActionsDetailsDialog.tsx
@@ -11,30 +11,30 @@ import {
   CircularProgress,
 } from "@mui/material";
 import CloseIcon from "@mui/icons-material/Close";
-import type { DetailsDialogProps } from "../../types";
+import type { DetailsDialogProps } from "../../shared/types";
 import { useTranslations } from "next-intl";
 import { useQuery } from "@tanstack/react-query";
-import { ProjectSharingTasksApi } from "@/services/api/projects/project-sharing-tasks";
+import { ProjectSharingProcedureApi } from "@/services/api/projects/project-sharing-procedure";
 
-const TasksDetailsDialog = ({
+const ActionsDetailsDialog = ({
   open,
   setOpenModal,
   rowId,
 }: DetailsDialogProps) => {
-  const tDetails = useTranslations("projectSettings.addTasks.details");
+  const tDetails = useTranslations("projectSettings.actions.details");
 
   const handleClose = () => setOpenModal(false);
 
   const detailQuery = useQuery({
-    queryKey: ["project-sharing-tasks", rowId],
+    queryKey: ["project-sharing-procedure", rowId],
     queryFn: async () => {
-      const res = await ProjectSharingTasksApi.show(rowId!);
+      const res = await ProjectSharingProcedureApi.show(rowId!);
       return res.data.payload;
     },
     enabled: open && Boolean(rowId),
   });
 
-  const task = detailQuery.data;
+  const action = detailQuery.data;
 
   return (
     <Dialog
@@ -76,23 +76,24 @@ const TasksDetailsDialog = ({
           <Box sx={{ display: "flex", justifyContent: "center", py: 4 }}>
             <CircularProgress />
           </Box>
-        ) : detailQuery.isError || !task ? (
+        ) : detailQuery.isError || !action ? (
           <Typography textAlign="center" color="error">
             {tDetails("notFound")}
           </Typography>
         ) : (
           <Box sx={{ display: "flex", flexDirection: "column", gap: 2, pt: 1 }}>
             <Typography variant="body2">
-              <strong>{tDetails("taskCode")}:</strong> {task.code}
+              <strong>{tDetails("actionCode")}:</strong> {action.code}
             </Typography>
             <Typography variant="body2">
-              <strong>{tDetails("taskName")}:</strong> {task.name}
+              <strong>{tDetails("actionDescription")}:</strong>{" "}
+              {action.description}
             </Typography>
             <Typography variant="body2" color="text.secondary">
-              <strong>{tDetails("createdAt")}:</strong> {task.created_at}
+              <strong>{tDetails("createdAt")}:</strong> {action.created_at}
             </Typography>
             <Typography variant="body2" color="text.secondary">
-              <strong>{tDetails("updatedAt")}:</strong> {task.updated_at}
+              <strong>{tDetails("updatedAt")}:</strong> {action.updated_at}
             </Typography>
           </Box>
         )}
@@ -101,4 +102,4 @@ const TasksDetailsDialog = ({
   );
 };
 
-export default TasksDetailsDialog;
+export default ActionsDetailsDialog;
