+++ modules/projects/settings/views/main-view/tab-views/work-orders/section/SectionView.tsx
@@ -1,14 +1,14 @@
 import { Box, Button, IconButton, Typography } from "@mui/material";
 import ArrowBackIcon from "@mui/icons-material/ArrowBack";
-import { CARDTYPE } from ".";
+import type { CARDTYPE } from "../card-types";
 import { Dispatch, SetStateAction, useMemo, useState } from "react";
 import { useTranslations } from "next-intl";
 import HeadlessTableLayout from "@/components/headless/table";
-import { RowActions } from "./components/RowActions";
-import EditSectionDialog from "./components/dialogs/EditSectionDialog";
-import SectionDetailsDialog from "./components/dialogs/SectionDetailsDialog";
-import type { ProjectSharingDepartment } from "./types";
-import AddSectionDialog from "./components/dialogs/AddSectionDialog";
+import { RowActions } from "../shared/RowActions";
+import EditSectionDialog from "./dialogs/EditSectionDialog";
+import SectionDetailsDialog from "./dialogs/SectionDetailsDialog";
+import type { ProjectSharingDepartment } from "../shared/types";
+import AddSectionDialog from "./dialogs/AddSectionDialog";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { ProjectSharingDepartmentApi } from "@/services/api/projects/project-sharing-department";
 import DeleteButton from "@/components/shared/delete-button";
