+++ modules/projects/settings/views/main-view/tab-views/work-orders/shared/RowActions.tsx
@@ -10,7 +10,7 @@ import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import VisibilityIcon from "@mui/icons-material/Visibility";
 import BorderColorIcon from "@mui/icons-material/BorderColor";
 import DeleteOutlineIcon from "@mui/icons-material/DeleteOutline";
-import { RowActionsProps } from "../types";
+import type { RowActionsProps } from "./types";
 import { useTranslations } from "next-intl";
 
 export function RowActions({
