+++ modules/projects/settings/views/main-view/tab-views/work-orders/tasks/TasksView.tsx
@@ -1,14 +1,14 @@
 import { Box, Button, IconButton, Typography } from "@mui/material";
 import ArrowBackIcon from "@mui/icons-material/ArrowBack";
-import { CARDTYPE } from ".";
+import type { CARDTYPE } from "../card-types";
 import { Dispatch, SetStateAction, useMemo, useState } from "react";
 import { useTranslations } from "next-intl";
 import HeadlessTableLayout from "@/components/headless/table";
-import { RowActions } from "./components/RowActions";
-import EditTasksDialog from "./components/dialogs/EditTasksDialog";
-import TasksDetailsDialog from "./components/dialogs/TasksDetailsDialog";
-import type { ProjectSharingTask } from "./types";
-import AddTasksDialog from "./components/dialogs/AddTasksDialog";
+import { RowActions } from "../shared/RowActions";
+import EditTasksDialog from "./dialogs/EditTasksDialog";
+import TasksDetailsDialog from "./dialogs/TasksDetailsDialog";
+import type { ProjectSharingTask } from "../shared/types";
+import AddTasksDialog from "./dialogs/AddTasksDialog";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { ProjectSharingTasksApi } from "@/services/api/projects/project-sharing-tasks";
 import DeleteButton from "@/components/shared/delete-button";
